autofill github.com/sapcc/go-api-declarations/bininfo variables
The shell invocations to get version, commit and build date are taken
from sapcc/limesctl, with some slight adjustments:

- `git describe` now has `--always` to work in repos without tags
  (such as Limes and Keppel).
- The timezone suffix in the date string is changed to match the
  `time.RFC3339` format (suffix `Z` instead of `UTC`).

In build rules, the `binName` variable is filled with the name of the
binary that is being compiled. During tests, the basename of the module
path is chosen as a sane fallback.