Dockerfile: do not use user "nobody"
Instead, create our own user account and group with well-defined static
names and IDs.

Also, we generate the Dockerfile's USER command with numeric IDs instead
of names now. This is a significant change because Kubernetes uses this
metadata to validate the `runAsNonRoot` security directive. If a named
user is given, it cannot be sure that the ID for that name is not 0, so
it then requires that the user ID be overriden with something non-zero
in the container spec (which introduces unnecessary coupling between the
container spec and the image contents).

The aforementioned change to the USER command would have required
rethinking what we accept for `dockerfile.user`. But since the only
practical usecase for `dockerfile.user` is to use `root` instead of a
non-privileged user (in swift-drive-autopilot), I replaced it with the
more specific `dockerfile.runAsRoot` option.