renovate config: set custom value for CommitMessageAction
I want to have a filter rule in my mail client that matches only
notifications for Renovate PRs, but not other GitHub notifications.
Because Renovatebot is not the only user participating in this PR,
I cannot filter on From. Realistically only the PR title can be used
as a basis for the rule.

Setting the PR title directly is supported in Renovate, but support is
deprecated. By default, the PR title is equal to the commit message.

Out of all fields in the commit message that can be customized, the
obvious choices are commitMessageAction (the verb at the start, "Update"
by default) or commitMessageExtra (a suffix at the end, empty by
default). I was going with the first option and setting
commitMessageAction to "Renovate" instead of "Update", but based on
feedback we're leaving "Update" in and putting "Renovate:" in front.